Dynamic form + changes ever since
This commit includes all of my changes ever since Summer 2022 in one commit.

I've made sure that the code works and formatted it correctly.

For what we need to do next, we need to finalize the Scouting App (add SQLlite and other miscellaneous things), then we will begin work on the MasterApp once this is done.